# backend/artifacts/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/ops/slo-snapshot": {
      "get": {
        "operationId": "snapshot",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aggregated technical and security metrics from PostgreSQL source of truth."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read operational SLO snapshot for dashboards",
        "tags": [
          "Operational readiness"
        ]
      }
    },
...
    "/v1/owner/appointments": {
      "get": {
        "operationId": "listAppointments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Только заявки и записи владельца из bearer JWT."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Список заявок и записей текущего владельца",
        "tags": [
          "Owner profile"
        ]
      }
    },
    "/v1/owner/bookings": {
      "get": {
        "operationId": "listBookings",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "petId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner profile"
        ]
      }
    },
    "/v1/owner/appointments/{holdId}": {
      "get": {
        "operationId": "appointmentDetail",
        "parameters": [
          {
            "name": "holdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статус, клиника, питомец, услуга, timeline и доступные действия."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подробный authoritative snapshot записи владельца",
        "tags": [
          "Owner profile"
        ]
      }
    },
    "/v1/owner/bookings/{holdId}": {
      "get": {
        "operationId": "bookingDetail",
        "parameters": [
          {
            "name": "holdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/appointments": {
      "get": {
        "operationId": "ClinicAppointmentsRegistry_list",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "bucket",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "upcoming",
                "history"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stable cursor page from a PostgreSQL-owned snapshot boundary.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "clinicId",
                    "locationId",
                    "serverNow",
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "clinicId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "locationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "serverNow": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "appointmentId",
                          "aggregateVersion",
                          "statusCode",
                          "statusLabel",
                          "slot",
                          "pet",
                          "service"
                        ],
                        "properties": {
                          "appointmentId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "aggregateVersion": {
                            "type": "integer"
                          },
                          "statusCode": {
                            "type": "string",
                            "enum": [
                              "SCHEDULED",
                              "COMPLETED",
                              "NO_SHOW",
                              "CANCELLED"
                            ]
                          },
                          "statusLabel": {
                            "type": "string"
                          },
                          "slot": {
                            "type": "object",
                            "required": [
                              "startsAt",
                              "endsAt"
                            ],
                            "properties": {
                              "startsAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "endsAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          },
                          "pet": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "speciesLabel"
                            ],
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "speciesLabel": {
                                "type": "string"
                              }
                            }
                          },
                          "service": {
                            "type": "object",
                            "nullable": true,
                            "required": [
                              "displayName"
                            ],
                            "properties": {
                              "displayName": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bucket, limit or cursor is invalid."
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Capability, clinic, location and active membership are required."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Location-scoped clinic appointments registry",
        "tags": [
          "Clinic Portal"
        ]
      }
    },
    "/v1/clinic/{clinicId}/locations/{locationId}/appointments/{appointmentId}": {
      "get": {
        "operationId": "ClinicAppointmentsRegistry_detail",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicAppointmentDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Appointment, capability, clinic, location and active membership are required."
          },
          "404": {
            "description": "Appointment registry rollout is disabled."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Location-scoped clinic appointment administrative detail",
        "tags": [
          "Clinic Portal"
        ]
      }
    },
...
    "/v1/clinics": {
      "get": {
        "operationId": "listClinics",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "radiusKm",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availableFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availableTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openNow",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "telemedAvailable",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emergencyCapability",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "selectedPetId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Активные клиники без внутренних MIS/provider полей."
          }
        },
        "summary": "Публичный список клиник",
        "tags": [
          "Public catalog"
        ]
      }
    },
    "/v1/clinics/{clinicId}": {
      "get": {
        "operationId": "readClinic",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Публичная карточка клиники",
        "tags": [
          "Public catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "VetHelp Booking Core API",
    "description": "Контракт локального удержания слотов, подтверждения записи клиникой и освобождения hold.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "opaque Owner session or legacy staff JWT",
        "type": "http",
        "description": "Opaque vh_ Owner session or compatibility JWT for staff/system roles. Authorization and expiry are evaluated server-side for Owner sessions."
      }
    },
    "schemas": {
...
      "ClinicAppointmentDetailAppointmentDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "string",
            "format": "uuid"
          },
          "aggregateVersion": {
            "type": "integer",
            "minimum": 0
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "NO_SHOW",
              "CANCELLED",
              "UNKNOWN"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "appointmentId",
          "aggregateVersion",
          "statusCode",
          "statusLabel",
          "createdAt"
        ]
      },
      "ClinicAppointmentDetailScheduleDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string"
          },
          "sourceLabel": {
            "type": "string"
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "timezone",
          "sourceLabel"
        ]
      },
      "ClinicAppointmentDetailDisplayNameDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ]
      },
      "ClinicAppointmentDetailPetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "speciesLabel": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "speciesLabel"
        ]
      },
      "ClinicAppointmentDetailDto": {
        "type": "object",
        "properties": {
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          },
          "appointment": {
            "$ref": "#/components/schemas/ClinicAppointmentDetailAppointmentDto"
          },
          "schedule": {
            "$ref": "#/components/schemas/ClinicAppointmentDetailScheduleDto"
          },
          "owner": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicAppointmentDetailDisplayNameDto"
              }
            ]
          },
          "pet": {
            "$ref": "#/components/schemas/ClinicAppointmentDetailPetDto"
          },
          "service": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicAppointmentDetailDisplayNameDto"
              }
            ]
          },
          "veterinarian": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicAppointmentDetailDisplayNameDto"
              }
            ]
          },
          "resource": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicAppointmentDetailDisplayNameDto"
              }
            ]
          },
          "availableActions": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "clinicId",
          "locationId",
          "serverNow",
          "appointment",
          "schedule",
          "owner",
          "pet",
          "service",
          "veterinarian",
          "resource",
          "availableActions"
        ]
      },
...
    }
  }
}